feat: add ECB currency conversion to portfolio rebalancing
- expose backend-managed base and supported currencies
- retain each asset's quote currency and convert prices with ECB rates
- cache FX rates with configurable TTL and maximum age
- migrate monetary calculations and rebalance optimization to Decimal
- require a portfolio currency across API, storage, and import/export
- add the currency selector and result summary to the responsive UI
- extend backend and frontend coverage for FX and precision handling

+++ app/schemas/result.py
@@ -1,5 +1,7 @@
 """Pydantic v2 schemas for the HTTP response boundary."""
 
+from decimal import Decimal
+
 from pydantic import BaseModel, field_serializer
 
 from app.core.formatting import truncate, truncate2
@@ -8,34 +10,34 @@ from app.core.formatting import truncate, truncate2
 class AssetResultOut(BaseModel):
     id: int
     ticker: str
-    current_percentage: float
-    desired_percentage: float
-    shares: float
-    allocated: float
-    ticker_price: float
-    fees: float
-    buy: float
+    current_percentage: Decimal
+    desired_percentage: Decimal
+    shares: Decimal
+    allocated: Decimal
+    ticker_price: Decimal
+    fees: Decimal
+    buy: Decimal
 
     @field_serializer(
         "current_percentage", "desired_percentage", "shares",
         "allocated", "ticker_price", "fees",
     )
-    def _fmt_fields(self, v: float) -> float:
-        return truncate2(v)
+    def _fmt_fields(self, v: Decimal) -> float:
+        return float(truncate2(v))
 
     @field_serializer("buy")
-    def _fmt_buy(self, v: float) -> float:
+    def _fmt_buy(self, v: Decimal) -> float:
         # Whole-share mode yields integer-valued floats (5.0); fractional mode
         # yields up to 6 decimals. 6 dp is finer than any broker and keeps the
         # residual unspent cash below a cent.
-        return truncate(v, 6)
+        return float(truncate(v, 6))
 
 
 class RebalanceResponse(BaseModel):
     results: list[AssetResultOut]
-    total_fees: float
-    change: float
+    total_fees: Decimal
+    change: Decimal
 
     @field_serializer("total_fees", "change")
-    def _fmt_totals(self, v: float) -> float:
-        return truncate2(v)
+    def _fmt_totals(self, v: Decimal) -> float:
+        return float(truncate2(v))
